Added utils module to contain various lpc and mfcc extraction functions.
